Micro-Services/game_app/requirements.txt
Django==5.0.1
djangorestframework
psycopg2
psycopg2-binary
django-cors-headers
django-prometheus
PyJWT
uvicorn